Add subject management page

+++ src/components/DeleteSubjectButton.tsx
@@ -1,18 +1,35 @@
 "use client";
 
-import { startTransition } from "react";
+import { useTransition } from "react";
 import { deleteSubject } from "../server-functions/deleteSubject";
+import { IconTrash } from "@tabler/icons-react";
+import { ActionIcon } from "@mantine/core";
 
-export function DeleteSubjectButton({ subjectId }: { subjectId: number }) {
+export function DeleteSubjectButton({
+  subjectId,
+  className,
+}: {
+  subjectId: number;
+  className?: string;
+}) {
+  const [pending, trackTransition] = useTransition();
+
   return (
-    <button
+    <ActionIcon
+      className={className}
+      loading={pending}
+      variant="subtle"
+      color="red"
+      size="lg"
       onClick={() => {
-        startTransition(async () => {
-          await deleteSubject(subjectId);
-        });
+        if (confirm("Opravdu smazat předmět a s ním všechny jeho dokumenty?")) {
+          trackTransition(async () => {
+            await deleteSubject(subjectId);
+          });
+        }
       }}
     >
-      Smazat předmět a jeho dokumenty
-    </button>
+      <IconTrash />
+    </ActionIcon>
   );
 }
